♻️ refactor : 스키마 분리  #252
