+++ src/components/auth/auth-form.tsx
@@ -10,7 +10,7 @@ import {
   duplicateEmailValidation,
   duplicateNickNameValidation,
 } from '@/utils/auth/duplicate-validation';
-import { loginSchema, signUpSchema } from '@/utils/auth/validate-schema';
+import { loginSchema, signUpSchema } from '@/utils/auth/auth-validate-schema';
 import { debounce } from '@/utils/common/common.debounce.utils';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useEffect, useMemo, useState } from 'react';
@@ -19,7 +19,6 @@ import AuthErrorIcon from '@/components/icons/auth/auth-error';
 import AuthConfirmIcon from '@/components/icons/auth/auth-confirm';
 import AuthKakaoIcon from '@/components/icons/auth/auth-kakao';
 import AuthGoogleIcon from '@/components/icons/auth/auth-google';
-import AuthXButtonIcon from '@/components/icons/auth/auth-x-button';
 
 interface FormType {
   email: string;
